Reapply "Flask Blueprints Implemented"
This reverts commit 61a81d7839481ec1c8a22cbc3817b9a59b9afd95.
