+++ .gitignore
@@ -6,3 +6,4 @@
 /bertimbau-finetuned
 /UX-Tracking Extension
 /UX-Tracking-Extension
+/*/__pycache__
