+++ external/blueprints.py
@@ -1,0 +1,4 @@
+from .auth import external_auth_bp
+from .receivedata import external_receivedata_bp
+
+external_bps = (external_auth_bp, external_receivedata_bp)
