+++ templates/erro.html
@@ -65,7 +65,7 @@
             <p class="apresentacao__conteudo__titulo display-6 fw-bold lh-1 mb-3">Erro {{erro}}</p>
             <p class="apresentacao__conteudo__texto lead">{{message}}</p>
             <P class="apresentacao__conteudo__texto lead">Se desejar obter mais informações ou relatar o erro, por favor, entre em contato conosco através do e-mail <a class="apresentacao__conteudo__link lead" href="mailto:uxtracking.service@gmail.com">uxtracking.service@gmail.com</a></P>
-            <p class="apresentacao__conteudo__texto lead">Você pode retornar à página inicial clicando no link: <a class="apresentacao__conteudo__link lead" href="{{url_for('index')}}">Voltar para Página Inicial</a></p>
+            <p class="apresentacao__conteudo__texto lead">Você pode retornar à página inicial clicando no link: <a class="apresentacao__conteudo__link lead" href="{{url_for('index_bp.index_get')}}">Voltar para Página Inicial</a></p>
         </section>
     </main>
 </body>
