+++ templates/forgot_pass.html
@@ -63,8 +63,8 @@
               <button class="btn btn-primary w-100 py-2" type="submit">Receber minha senha por email</button>
 
               <p style="text-align: center;"><br>
-                <a href="{{ url_for('login') }}">Tem uma conta?</a><br>
-                <a href="{{ url_for('register') }}">Não tem uma conta?</a>
+                <a href="{{ url_for('auth_bp.login_get') }}">Tem uma conta?</a><br>
+                <a href="{{ url_for('auth_bp.register_get') }}">Não tem uma conta?</a>
               </p>
             </form>
           </div>
