+++ templates/publicacoes.html
@@ -64,7 +64,7 @@
           
           .content:hover {
             transform: scale(1.05); /* Aumenta ligeiramente ao passar o mouse */
-            color: #fffff; /* Opcional: muda a cor do texto ao passar o mouse */
+            color: #ffffff; /* Opcional: muda a cor do texto ao passar o mouse */
             background: #f57307;
           }
