+++ templates/register.html
@@ -137,7 +137,7 @@
               <h1 class="h4 mb-3 fw-normal" style="text-align: center;">ou</h1>
               
               <div style="text-align:center">
-                <a href="{{ url_for('google') }}">
+                <a href="{{ url_for('auth_bp.google') }}">
                   <button type="button" class="login-with-google-btn">
                     Registre-se com o Google
                   </button>
@@ -145,8 +145,8 @@
               </div>
             -->
               <p style="text-align: center;"><br>
-                <a href="{{ url_for('forgot_pass') }}">Esqueceu sua senha?</a><br>
-                <a href="{{ url_for('login') }}">Já tem conta?</a>
+                <a href="{{ url_for('pass_bp.forgot_pass_get') }}">Esqueceu sua senha?</a><br>
+                <a href="{{ url_for('auth_bp.login_get') }}">Já tem conta?</a>
               </p>
             </form>
         </div>
